requirements.txt
python-telegram-bot==20.7
httpx==0.26.0
apscheduler==3.10.4
